Address review feedback on the quick scan comment field
Move the quick-scan field resolution out of the handler and into the service
layer: ResolveQuickScanFields is now a ReceiptService method in
services/quick_scan_fields.go, leaving the handler to parse the request and
shape the response. Its repositories and PermissionService are built from the
service's transaction rather than nil, matching GetReceiptForUser.

quick_scan_resolve_test.go moves to the services package with it. The comment
tests stay in handlers -- three of them go through the HTTP handler and all
eight share one seed helper, so splitting the file would duplicate the seed for
no gain; the resolver cases just call the exported method.

Count the comment limit in runes rather than bytes. The Comment column is
varchar(500), which MySQL and Postgres both measure in characters, so len() was
stricter than the column it exists to mirror and rejected an accented or
non-Latin comment well inside its real capacity -- while mobile's 500-character
cap let it through, so client and server disagreed. The new boundary test was
confirmed to fail against the old check before the fix landed. The limit is now
published on the contract as maxLength: 500 (OpenAPI measures characters too);
it is doc-only for both generators, so neither client changed.

Cap the comment at 500 on the desktop as well, so an over-length comment is
caught before it can come back as a 400 after submit. The validator is attached
where the control is created, not in the per-image recompute, because
setRequired composes validators additively and would otherwise re-add it every
pass. base-input has no message for the maxlength error and would have rendered
an empty mat-error, so the textarea also supplies additionalErrorMessages.

Two review comments are not acted on, each answered in full on its thread: the
duplicate `const requests` report is a false positive (the two declarations are
in separate test blocks, and the spec compiles and runs green), and the mobile
Key-based finder request contradicts the documented convention for
FormBuilderTextField, which has no Key and is matched by its name.
